Move commands, menus, listeners to interaction
This is a top level package that contains all ingame interactions.
