+++ src/main/kotlin/dev/mizarc/waystonewarps/interaction/commands/WarpMenuCommand.kt
@@ -1,4 +1,4 @@
-package dev.mizarc.waystonewarps.commands
+package dev.mizarc.waystonewarps.interaction.commands
 
 import co.aikar.commands.BaseCommand
 import co.aikar.commands.annotation.CommandAlias;
@@ -7,7 +7,7 @@ import co.aikar.commands.annotation.Default;
 import co.aikar.commands.annotation.Dependency;
 import org.bukkit.entity.Player
 import dev.mizarc.waystonewarps.Teleporter
-import dev.mizarc.waystonewarps.menus.WarpMenu
+import dev.mizarc.waystonewarps.interaction.menus.WarpMenu
 
 @CommandAlias("warpmenu")
 @CommandPermission("worldwidewarps.command.warpmenu")
