+++ src/main/kotlin/dev/mizarc/waystonewarps/interaction/listeners/PlayerRegistrationListener.kt
@@ -1,4 +1,4 @@
-package dev.mizarc.waystonewarps.listeners
+package dev.mizarc.waystonewarps.interaction.listeners
 
 import net.milkbowl.vault.chat.Chat
 import org.bukkit.event.EventHandler
