+++ src/main/kotlin/dev/mizarc/waystonewarps/interaction/listeners/WarpMoveToolRemovalListener.kt
@@ -1,4 +1,4 @@
-package dev.mizarc.waystonewarps.listeners
+package dev.mizarc.waystonewarps.interaction.listeners
 
 import dev.mizarc.waystonewarps.utils.getStringMeta
 import org.bukkit.Material
